Give each command a parent, add FullName method
The Parent value on the command struct will be a pointer to their parent
if they have one. This will be used to generating the command's full
name if it is a sub-command.

For example given the command 'add' as a child of the command 'remote'
the full name would be 'remote-add'.