Add support for multiple passings of the same flag
The same flag can now be passed multiple times to a single command. The
flag Get methods by default will return the first occurrence of the given
flag. GetAll can be used to get all occurrences of the given flag.

If a flag with a handler is passed multiple times then the handler will
only be called once, from the first occurrence of that flag.